@@ -13,7 +13,6 @@ from rest_framework.response import Response
 from rest_framework.views import APIView
 from rest_framework_jsonp.renderers import JSONPRenderer
 
-from ..constants import DECISIONS
 from ..decisionengine import get_ad_decision_backend
 from ..models import AdImpression
 from ..models import Advertisement
@@ -174,8 +173,6 @@ class AdDecisionView(GeoIpMixin, APIView):
                     url=url,
                 )
                 return {}
-
-            ad.incr(impression_type=DECISIONS, publisher=publisher)
 
             data = ad.offer_ad(
                 request=self.request,
@@ -356,7 +353,7 @@ class AdvertiserViewSet(viewsets.ReadOnlyModelViewSet):
         return self.request.user.advertisers.all()
 
     @action(detail=True, methods=["get"])
-    def report(self, request, slug=None):
+    def report(self, request, slug=None):  # pylint: disable=unused-argument
         """Return a report of ad performance for this advertiser."""
         # This will raise a 404 if the user doesn't have access to the advertiser
         advertiser = self.get_object()
@@ -465,7 +462,7 @@ class PublisherViewSet(viewsets.ReadOnlyModelViewSet):
         return self.request.user.publishers.all()
 
     @action(detail=True, methods=["get"])
-    def report(self, request, slug=None):
+    def report(self, request, slug=None):  # pylint: disable=unused-argument
         """Return a report of ad performance for this publisher."""
         # This will raise a 404 if the user doesn't have access to the publisher
         publisher = self.get_object()
