@@ -7,6 +7,7 @@ import uuid
 from collections import Counter
 
 import bleach
+import djstripe.models as djstripe_models
 from django.conf import settings
 from django.core.validators import MaxValueValidator
 from django.core.validators import MinValueValidator
@@ -19,12 +20,13 @@ from django.template.loader import get_template
 from django.urls import reverse
 from django.utils import timezone
 from django.utils.crypto import get_random_string
+from django.utils.functional import cached_property
 from django.utils.html import mark_safe
 from django.utils.text import slugify
 from django.utils.translation import ugettext_lazy as _
-from django_countries import countries
 from django_countries.fields import CountryField
 from django_extensions.db.models import TimeStampedModel
+from djstripe.enums import InvoiceStatus
 from jsonfield import JSONField
 from simple_history.models import HistoricalRecords
 from user_agents import parse
@@ -39,12 +41,16 @@ from .constants import IMPRESSION_TYPES
 from .constants import OFFERS
 from .constants import PAID
 from .constants import PAID_CAMPAIGN
+from .constants import PAYOUT_OPENCOLLECTIVE
+from .constants import PAYOUT_PAYPAL
 from .constants import PAYOUT_STATUS
+from .constants import PAYOUT_STRIPE
 from .constants import PENDING
 from .constants import PUBLISHER_PAYOUT_METHODS
 from .constants import VIEWS
 from .utils import anonymize_ip_address
 from .utils import calculate_ctr
+from .utils import COUNTRY_DICT
 from .utils import generate_absolute_url
 from .utils import get_ad_day
 from .utils import get_client_country
@@ -125,6 +131,17 @@ class Publisher(TimeStampedModel, IndestructibleModel):
             "Whether this publisher allows unauthenticated ad decision API requests (eg. JSONP)"
         ),
     )
+    disabled = models.BooleanField(
+        default=False,
+        help_text=_("Completely disable this publisher"),
+    )
+
+    saas = models.BooleanField(
+        default=False,
+        help_text=_(
+            "This published is configured as a SaaS customer. They will be billed by usage instead of paid out."
+        ),
+    )
 
     # Default to False so that we can use this as an "approved" flag for publishers
     allow_paid_campaigns = models.BooleanField(_("Allow paid campaigns"), default=False)
@@ -157,6 +174,16 @@ class Publisher(TimeStampedModel, IndestructibleModel):
         null=True,
         default=None,
     )
+    djstripe_account = models.ForeignKey(
+        djstripe_models.Account,
+        verbose_name=_("Stripe connected account"),
+        on_delete=models.SET_NULL,
+        blank=True,
+        null=True,
+        default=None,
+    )
+
+    # Deprecated - migrate to `stripe_account`
     stripe_connected_account_id = models.CharField(
         _("Stripe connected account ID"),
         max_length=200,
@@ -253,11 +280,11 @@ class Publisher(TimeStampedModel, IndestructibleModel):
         return 0
 
     def payout_url(self):
-        if self.stripe_connected_account_id:
-            return f"https://dashboard.stripe.com/connect/accounts/{self.stripe_connected_account_id}"
-        if self.open_collective_name:
+        if self.payout_method == PAYOUT_STRIPE and self.djstripe_account.id:
+            return f"https://dashboard.stripe.com/connect/accounts/{self.djstripe_account.id}"
+        if self.payout_method == PAYOUT_OPENCOLLECTIVE and self.open_collective_name:
             return f"https://opencollective.com/{self.open_collective_name}"
-        if self.paypal_email:
+        if self.payout_method == PAYOUT_PAYPAL and self.paypal_email:
             return "https://www.paypal.com/myaccount/transfer/homepage/pay"
         return ""
 
@@ -295,6 +322,16 @@ class Advertiser(TimeStampedModel, IndestructibleModel):
     name = models.CharField(_("Name"), max_length=200)
     slug = models.SlugField(_("Advertiser Slug"), max_length=200, unique=True)
 
+    djstripe_customer = models.ForeignKey(
+        djstripe_models.Customer,
+        verbose_name=_("Stripe Customer"),
+        on_delete=models.SET_NULL,
+        blank=True,
+        null=True,
+        default=None,
+    )
+
+    # Deprecated - will migration to `customer`
     stripe_customer_id = models.CharField(
         _("Stripe Customer ID"), max_length=200, blank=True, null=True, default=None
     )
@@ -396,7 +433,7 @@ class Campaign(TimeStampedModel, IndestructibleModel):
         )
         if exclude_deprecated:
             queryset = queryset.exclude(deprecated=True)
-        return queryset
+        return queryset.distinct()
 
     def total_value(self):
         """Calculate total cost/revenue for all ads/flights in this campaign."""
@@ -498,6 +535,16 @@ class Flight(TimeStampedModel, IndestructibleModel):
         default=0, help_text=_("Clicks across all ads in this flight")
     )
 
+    # Connect to Stripe invoice data
+    # There can be multiple invoices for a flight
+    # (say a 3 month flight billed monthly)
+    # and an invoice can cover multiple flights
+    invoices = models.ManyToManyField(
+        djstripe_models.Invoice,
+        verbose_name=_("Stripe invoices"),
+        blank=True,
+    )
+
     history = HistoricalRecords()
 
     class Meta:
@@ -563,15 +610,13 @@ class Flight(TimeStampedModel, IndestructibleModel):
 
     def get_include_countries_display(self):
         included_country_codes = self.included_countries
-        countries_dict = dict(countries)
-        return [countries_dict.get(cc, "Unknown") for cc in included_country_codes]
+        return [COUNTRY_DICT.get(cc, "Unknown") for cc in included_country_codes]
 
     def get_exclude_countries_display(self):
         excluded_country_codes = self.excluded_countries
-        countries_dict = dict(countries)
-        return [countries_dict.get(cc, "Unknown") for cc in excluded_country_codes]
+        return [COUNTRY_DICT.get(cc, "Unknown") for cc in excluded_country_codes]
 
-    def show_to_geo(self, country_code, state_province_code=None, metro_code=None):
+    def show_to_geo(self, geo_data):
         """
         Check if a flight is valid for a given country code.
 
@@ -579,16 +624,19 @@ class Flight(TimeStampedModel, IndestructibleModel):
         will not match a flight with any ``include_countries`` but wont be
         excluded from any ``exclude_countries``
         """
-        if self.included_countries and country_code not in self.included_countries:
+        if self.included_countries and geo_data.country not in self.included_countries:
             return False
         if (
             self.included_state_provinces
-            and state_province_code not in self.included_state_provinces
+            and geo_data.region not in self.included_state_provinces
         ):
             return False
-        if self.included_metro_codes and metro_code not in self.included_metro_codes:
+        if (
+            self.included_metro_codes
+            and geo_data.metro not in self.included_metro_codes
+        ):
             return False
-        if self.excluded_countries and country_code in self.excluded_countries:
+        if self.excluded_countries and geo_data.country in self.excluded_countries:
             return False
 
         return True
@@ -768,6 +816,11 @@ class Flight(TimeStampedModel, IndestructibleModel):
         clicks = self.total_clicks
         views = self.total_views
         return calculate_ctr(clicks, views)
+
+    @cached_property
+    def active_invoices(self):
+        """Get invoices excluding drafts, void, and uncollectable ones."""
+        return self.invoices.filter(status__in=(InvoiceStatus.open, InvoiceStatus.paid))
 
 
 class AdType(TimeStampedModel, models.Model):
@@ -955,7 +1008,7 @@ class Advertisement(TimeStampedModel, IndestructibleModel):
 
     def __copy__(self):
         """Duplicate an ad."""
-        # https://docs.djangoproject.com/en/2.2/topics/db/queries/#copying-model-instances
+        # https://docs.djangoproject.com/en/3.2/topics/db/queries/#copying-model-instances
         # Get a fresh reference so that "self" doesn't become the new copy
         ad = Advertisement.objects.get(pk=self.pk)
 
@@ -996,25 +1049,40 @@ class Advertisement(TimeStampedModel, IndestructibleModel):
         TODO: Refactor this method, moving it off the Advertisement class since it can be called
               without an advertisement when we have a Decision and no Offer.
         """
-        assert impression_type in IMPRESSION_TYPES
         day = get_ad_day().date()
 
+        if isinstance(impression_type, str):
+            impression_types = (impression_type,)
+        else:
+            impression_types = impression_type
+
+        for imp_type in impression_types:
+            assert imp_type in IMPRESSION_TYPES
+
+            # Update the denormalized fields on the Flight
+            if imp_type == VIEWS:
+                Flight.objects.filter(pk=self.flight_id).update(
+                    total_views=models.F("total_views") + 1
+                )
+            elif imp_type == CLICKS:
+                Flight.objects.filter(pk=self.flight_id).update(
+                    total_clicks=models.F("total_clicks") + 1
+                )
+
         # Ensure that an impression object exists for today
-        impression, _ = AdImpression.objects.get_or_create(
-            advertisement=self, publisher=publisher, date=day
-        )
-        AdImpression.objects.filter(pk=impression.pk).update(
-            **{impression_type: models.F(impression_type) + 1}
+        # and make sure to query the writable DB for this
+        impression, created = AdImpression.objects.using("default").get_or_create(
+            advertisement=self,
+            publisher=publisher,
+            date=day,
+            defaults={imp_type: 1 for imp_type in impression_types},
         )
 
-        # Update the denormalized fields on the Flight
-        if impression_type == VIEWS:
-            Flight.objects.filter(pk=self.flight_id).update(
-                total_views=models.F("total_views") + 1
-            )
-        elif impression_type == CLICKS:
-            Flight.objects.filter(pk=self.flight_id).update(
-                total_clicks=models.F("total_clicks") + 1
+        if not created:
+            # If the object was created above, we don't need to update
+            # since the defaults will have already done the update for us.
+            AdImpression.objects.using("default").filter(pk=impression.pk).update(
+                **{imp_type: models.F(imp_type) + 1 for imp_type in impression_types}
             )
 
     def _record_base(
@@ -1030,13 +1098,19 @@ class Advertisement(TimeStampedModel, IndestructibleModel):
         user_agent = get_client_user_agent(request)
         client_id = get_client_id(request)
         parsed_ua = parse(user_agent)
-        country = get_client_country(request, ip_address)
+        country = get_client_country(request)
         url = url or request.META.get("HTTP_REFERER")
 
         if model != Click and settings.ADSERVER_DO_NOT_TRACK:
             # For compliance with DNT,
             # we can't store UAs indefinitely from a user merely browsing
             user_agent = None
+
+        if div_id:
+            # Even though the publisher could have a div of any length
+            # and we want to echo back the same div to them,
+            # we only store the first 100 characters of it.
+            div_id = div_id[: Offer.DIV_MAXLENGTH]
 
         obj = model.objects.create(
             date=timezone.now(),
@@ -1140,7 +1214,7 @@ class Advertisement(TimeStampedModel, IndestructibleModel):
         """
         ad_type = AdType.objects.filter(slug=ad_type_slug).first()
 
-        self.incr(impression_type=OFFERS, publisher=publisher)
+        self.incr(impression_type=(OFFERS, DECISIONS), publisher=publisher)
         offer = self._record_base(
             request=request,
             model=Offer,
@@ -1151,9 +1225,10 @@ class Advertisement(TimeStampedModel, IndestructibleModel):
             ad_type_slug=ad_type_slug,
         )
 
-        if forced:
-            # Ad offers forced to a specific ad or campaign should never be billed
-            # By discarding the nonce, the ad view/click will never count
+        if forced and self.flight.campaign.campaign_type == PAID_CAMPAIGN:
+            # Ad offers forced to a specific ad or campaign should never be billed.
+            # By discarding the nonce, the ad view/click will never count.
+            # We will still record data for unpaid campaign in reporting though.
             nonce = "forced"
         else:
             nonce = offer.pk
@@ -1622,6 +1697,8 @@ class AdBase(TimeStampedModel, IndestructibleModel):
 
     """A base class for data on ad views and clicks."""
 
+    DIV_MAXLENGTH = 100
+
     date = models.DateTimeField(_("Impression date"), db_index=True)
 
     publisher = models.ForeignKey(
@@ -1660,7 +1737,9 @@ class AdBase(TimeStampedModel, IndestructibleModel):
 
     # Client data
     keywords = JSONField(_("Keyword targeting for this view"), blank=True, null=True)
-    div_id = models.CharField(_("Div id"), blank=True, null=True, max_length=100)
+    div_id = models.CharField(
+        _("Div id"), blank=True, null=True, max_length=DIV_MAXLENGTH
+    )
     # This locked up the DB for a long time trying to write to our huge View table,
     # so we made it a Text field instead of a FK.
     ad_type_slug = models.CharField(_("Ad type"), blank=True, null=True, max_length=100)
@@ -1789,6 +1868,8 @@ class Offer(AdBase):
     class Meta:
         # This is needed because we can't sort on pk to get the created ordering
         ordering = ("-date",)
+        if settings.ADSERVER_OFFER_DB_TABLE:
+            db_table = settings.ADSERVER_OFFER_DB_TABLE
 
 
 class PublisherPayout(TimeStampedModel):
