@@ -10,15 +10,12 @@ from django.conf import settings
 from django.core import management
 from django.db.models import Count
 from django.db.models import F
+from django.db.models import Q
 from django.utils.timezone import is_naive
 from django.utils.timezone import utc
 from django_slack import slack_message
 
-from .constants import CLICKS
-from .constants import IMPRESSION_TYPES
-from .constants import OFFERS
 from .constants import PAID_CAMPAIGN
-from .constants import VIEWS
 from .models import AdImpression
 from .models import Flight
 from .models import GeoImpression
@@ -29,20 +26,15 @@ from .models import Publisher
 from .models import RegionImpression
 from .models import RegionTopicImpression
 from .models import UpliftImpression
-from .regiontopics import africa
 from .regiontopics import backend_web
 from .regiontopics import blockchain
 from .regiontopics import data_science
 from .regiontopics import devops
-from .regiontopics import eu_aus_nz
-from .regiontopics import exclude
 from .regiontopics import frontend_web
 from .regiontopics import game_dev
-from .regiontopics import latin_america
+from .regiontopics import get_region_from_country_code
 from .regiontopics import python
 from .regiontopics import security_privacy
-from .regiontopics import us_ca
-from .regiontopics import wider_apac
 from .reports import PublisherReport
 from .utils import calculate_percent_diff
 from .utils import generate_absolute_url
@@ -68,30 +60,8 @@ def _get_day(day):
     return (start_date, end_date)
 
 
-def _default_filters(impression_type, start_date, end_date):
-    """Filter the queryset by date and impression type."""
-    # Use the replica for this query, since it's how we do all our reporting queries,
-    # and it currently hammers the prod DB.
-    queryset = Offer.objects.using(settings.REPLICA_SLUG).filter(
-        date__gte=start_date,
-        date__lt=end_date,  # Things at UTC midnight should count towards tomorrow
-        # is_refunded=False,  # This causes the query to be a filtered index and is much slower
-    )
-
-    if impression_type == CLICKS:
-        queryset = queryset.filter(clicked=True)
-    elif impression_type == VIEWS:
-        queryset = queryset.filter(viewed=True)
-    elif impression_type == OFFERS:
-        queryset = queryset.filter(advertisement__isnull=False)
-
-    return queryset
-
-
 @app.task()
-def daily_update_geos(
-    day=None, geo=True, region=True
-):  # pylint: disable=too-many-branches
+def daily_update_geos(day=None, geo=True, region=True):
     """
     Update the Geo & region index each day.
 
@@ -115,68 +85,81 @@ def daily_update_geos(
             date__lt=end_date,  # Things at UTC midnight should count towards tomorrow
         ).delete()
 
-    for impression_type in IMPRESSION_TYPES:
-        topic_mapping = defaultdict(int)
-        queryset = _default_filters(impression_type, start_date, end_date)
-        for values in (
-            queryset.values("advertisement", "country", "publisher")
-            .annotate(total=Count("country"))
-            .filter(total__gt=0)
-            .order_by("-total")
-        ):
-            country = values["country"]
-            if geo:
-                impression, _ = GeoImpression.objects.using("default").get_or_create(
-                    publisher_id=values["publisher"],
-                    advertisement_id=values["advertisement"],
-                    country=country,
-                    date=start_date,
-                )
-                GeoImpression.objects.using("default").filter(pk=impression.pk).update(
-                    **{impression_type: values["total"]}
-                )
-
-            if region:
-
-                if country in us_ca:
-                    _region = "us-ca"
-                elif country in eu_aus_nz:
-                    _region = "eu-aus-nz"
-                elif country in wider_apac:
-                    _region = "wider-apac"
-                elif country in latin_america:
-                    _region = "latin-america"
-                elif country in africa:
-                    _region = "africa"
-                elif country in exclude:
-                    _region = "exclude"
-                else:
-                    _region = "global"
-
-                publisher = values["publisher"]
-                advertisement = values["advertisement"]
-                topic_mapping[f"{advertisement}:{publisher}:{_region}"] += values[
-                    "total"
-                ]
+    topic_mapping = defaultdict(
+        lambda: {
+            "decisions": 0,
+            "offers": 0,
+            "views": 0,
+            "clicks": 0,
+        }
+    )
+    queryset = Offer.objects.using(settings.REPLICA_SLUG).filter(
+        date__gte=start_date,
+        date__lt=end_date,  # Things at UTC midnight should count towards tomorrow
+    )
+    for values in (
+        queryset.values("advertisement", "country", "publisher")
+        .annotate(
+            total_decisions=Count("country"),
+            total_offers=Count("country", filter=Q(advertisement__isnull=False)),
+            total_views=Count("country", filter=Q(viewed=True)),
+            total_clicks=Count("country", filter=Q(clicked=True)),
+        )
+        .filter(total_decisions__gt=0)
+        .order_by("-total_decisions")
+        .iterator()
+    ):
+        country = values["country"]
+        if geo:
+            impression, _ = GeoImpression.objects.using("default").get_or_create(
+                publisher_id=values["publisher"],
+                advertisement_id=values["advertisement"],
+                country=country,
+                date=start_date,
+            )
+            GeoImpression.objects.using("default").filter(pk=impression.pk).update(
+                decisions=values["total_decisions"],
+                offers=values["total_offers"],
+                views=values["total_views"],
+                clicks=values["total_clicks"],
+            )
 
         if region:
-            log.info(
-                "Saving %s RegionImpressions: %s", len(topic_mapping), impression_type
+            _region = get_region_from_country_code(country)
+
+            publisher = values["publisher"]
+            advertisement = values["advertisement"]
+            topic_mapping[f"{advertisement}:{publisher}:{_region}"][
+                "decisions"
+            ] += values["total_decisions"]
+            topic_mapping[f"{advertisement}:{publisher}:{_region}"]["offers"] += values[
+                "total_offers"
+            ]
+            topic_mapping[f"{advertisement}:{publisher}:{_region}"]["views"] += values[
+                "total_views"
+            ]
+            topic_mapping[f"{advertisement}:{publisher}:{_region}"]["clicks"] += values[
+                "total_clicks"
+            ]
+
+    if region:
+        for data, value in topic_mapping.items():
+            ad, publisher, _region = data.split(":")
+            # Handle the conversion of None
+            if ad == "None":
+                ad = None
+            impression, _ = RegionImpression.objects.using("default").get_or_create(
+                publisher_id=publisher,
+                advertisement_id=ad,
+                region=_region,
+                date=start_date,
             )
-            for data, value in topic_mapping.items():
-                ad, publisher, _region = data.split(":")
-                # Handle the conversion of None
-                if ad == "None":
-                    ad = None
-                impression, _ = RegionImpression.objects.using("default").get_or_create(
-                    publisher_id=publisher,
-                    advertisement_id=ad,
-                    region=_region,
-                    date=start_date,
-                )
-                RegionImpression.objects.using("default").filter(
-                    pk=impression.pk
-                ).update(**{impression_type: F(impression_type) + value})
+            RegionImpression.objects.using("default").filter(pk=impression.pk).update(
+                decisions=F("decisions") + value["decisions"],
+                offers=F("offers") + value["offers"],
+                views=F("views") + value["views"],
+                clicks=F("clicks") + value["clicks"],
+            )
 
 
 @app.task()
@@ -190,27 +173,39 @@ def daily_update_placements(day=None):
 
     log.info("Updating PlacementImpressions for %s-%s", start_date, end_date)
 
-    for impression_type in IMPRESSION_TYPES:
-        queryset = _default_filters(impression_type, start_date, end_date)
-        for values in (
-            queryset.values("publisher", "advertisement", "div_id", "ad_type_slug")
-            .annotate(total=Count("div_id"))
-            .filter(total__gt=0)
-            .filter(publisher__record_placements=True)
-            .exclude(div_id__regex=r"(rtd-\w{4}|ad_\w{4}).*")
-            .order_by("-total")
-        ):
+    queryset = Offer.objects.using(settings.REPLICA_SLUG).filter(
+        date__gte=start_date,
+        date__lt=end_date,  # Things at UTC midnight should count towards tomorrow
+    )
 
-            impression, _ = PlacementImpression.objects.using("default").get_or_create(
-                publisher_id=values["publisher"],
-                advertisement_id=values["advertisement"],
-                div_id=values["div_id"],
-                ad_type_slug=values["ad_type_slug"],
-                date=start_date,
-            )
-            PlacementImpression.objects.using("default").filter(
-                pk=impression.pk
-            ).update(**{impression_type: values["total"]})
+    for values in (
+        queryset.values("publisher", "advertisement", "div_id", "ad_type_slug")
+        .annotate(
+            total_decisions=Count("div_id"),
+            total_offers=Count("div_id", filter=Q(advertisement__isnull=False)),
+            total_views=Count("div_id", filter=Q(viewed=True)),
+            total_clicks=Count("div_id", filter=Q(clicked=True)),
+        )
+        .filter(total_decisions__gt=0)
+        .filter(publisher__record_placements=True)
+        .exclude(div_id__regex=r"(rtd-\w{4}|ad_\w{4}).*")
+        .order_by("-total_decisions")
+        .iterator()
+    ):
+
+        impression, _ = PlacementImpression.objects.using("default").get_or_create(
+            publisher_id=values["publisher"],
+            advertisement_id=values["advertisement"],
+            div_id=values["div_id"],
+            ad_type_slug=values["ad_type_slug"],
+            date=start_date,
+        )
+        PlacementImpression.objects.using("default").filter(pk=impression.pk).update(
+            decisions=values["total_decisions"],
+            offers=values["total_offers"],
+            views=values["total_views"],
+            clicks=values["total_clicks"],
+        )
 
 
 @app.task()
@@ -224,25 +219,36 @@ def daily_update_impressions(day=None):
 
     log.info("Updating AdImpressions for %s-%s", start_date, end_date)
 
-    for impression_type in IMPRESSION_TYPES:
-        queryset = _default_filters(impression_type, start_date, end_date)
+    queryset = Offer.objects.using(settings.REPLICA_SLUG).filter(
+        date__gte=start_date,
+        date__lt=end_date,  # Things at UTC midnight should count towards tomorrow
+    )
 
-        for values in (
-            queryset.values("publisher", "advertisement")
-            # This needs to be publisher and not advertisement to gets decisions properly
-            .annotate(total=Count("publisher"))
-            .filter(total__gt=0)
-            .order_by("-total")
-        ):
+    for values in (
+        queryset.values("publisher", "advertisement")
+        # This needs to be publisher and not advertisement to gets decisions properly
+        .annotate(
+            total_decisions=Count("publisher"),
+            total_offers=Count("publisher", filter=Q(advertisement__isnull=False)),
+            total_views=Count("publisher", filter=Q(viewed=True)),
+            total_clicks=Count("publisher", filter=Q(clicked=True)),
+        )
+        .filter(total_decisions__gt=0)
+        .order_by("-total_decisions")
+        .iterator()
+    ):
 
-            impression, _ = AdImpression.objects.using("default").get_or_create(
-                publisher_id=values["publisher"],
-                advertisement_id=values["advertisement"],
-                date=start_date,
-            )
-            AdImpression.objects.using("default").filter(pk=impression.pk).update(
-                **{impression_type: values["total"]}
-            )
+        impression, _ = AdImpression.objects.using("default").get_or_create(
+            publisher_id=values["publisher"],
+            advertisement_id=values["advertisement"],
+            date=start_date,
+        )
+        AdImpression.objects.using("default").filter(pk=impression.pk).update(
+            decisions=values["total_decisions"],
+            offers=values["total_offers"],
+            views=values["total_views"],
+            clicks=values["total_clicks"],
+        )
 
 
 @app.task()
@@ -262,52 +268,61 @@ def daily_update_keywords(day=None):
         date__lt=end_date,  # Things at UTC midnight should count towards tomorrow
     ).delete()
 
-    for impression_type in IMPRESSION_TYPES:
-        queryset = _default_filters(impression_type, start_date, end_date)
+    queryset = Offer.objects.using(settings.REPLICA_SLUG).filter(
+        date__gte=start_date,
+        date__lt=end_date,  # Things at UTC midnight should count towards tomorrow
+    )
 
-        for values in (
-            queryset.values("publisher", "advertisement", "keywords")
-            .annotate(total=Count("keywords"))
-            .filter(total__gt=0)
-            .order_by("-total")
-            .values(
-                "publisher",
-                "advertisement",
-                "keywords",
-                "advertisement__flight__targeting_parameters",
-                "total",
-            )
+    for values in (
+        queryset.values("publisher", "advertisement", "keywords", "viewed", "clicked")
+        .annotate(
+            total_decisions=Count("keywords"),
+            total_offers=Count("keywords", filter=Q(advertisement__isnull=False)),
+            total_views=Count("keywords", filter=Q(viewed=True)),
+            total_clicks=Count("keywords", filter=Q(clicked=True)),
+        )
+        .order_by("-total_decisions")
+        .values(
+            "publisher",
+            "advertisement",
+            "keywords",
+            "advertisement__flight__targeting_parameters",
+            "total_decisions",
+            "total_offers",
+            "total_views",
+            "total_clicks",
+        )
+        .iterator()
+    ):
+        if not (
+            values["keywords"] and values["advertisement__flight__targeting_parameters"]
         ):
-            if not (
-                values["keywords"]
-                and values["advertisement__flight__targeting_parameters"]
-            ):
-                continue
+            continue
 
-            keywords = json.loads(values["keywords"])
-            publisher_keywords = set(keywords)
+        publisher_keywords = set(values["keywords"])
+        flight_targeting = values["advertisement__flight__targeting_parameters"]
+        flight_keywords = set(flight_targeting.get("include_keywords", {}))
 
-            flight_targeting = json.loads(
-                values["advertisement__flight__targeting_parameters"]
+        # Only store keywords where the advertiser targeting
+        # matched the keywords on the offer
+        matched_keywords = publisher_keywords & flight_keywords
+
+        for keyword in matched_keywords:
+            impression, _ = KeywordImpression.objects.using("default").get_or_create(
+                date=start_date,
+                publisher_id=values["publisher"],
+                advertisement_id=values["advertisement"],
+                keyword=keyword,
             )
-            flight_keywords = set(flight_targeting.get("include_keywords", {}))
-
-            matched_keywords = publisher_keywords & flight_keywords
-            for keyword in matched_keywords:
-                impression, _ = KeywordImpression.objects.using(
-                    "default"
-                ).get_or_create(
-                    date=start_date,
-                    publisher_id=values["publisher"],
-                    advertisement_id=values["advertisement"],
-                    keyword=keyword,
-                )
-                # These are a Sum because we can't query for specific keywords from Postgres,
-                # so a specific publisher and advertisement set could return the same keyword:
-                # ['python', 'django'] and ['python, 'flask'] both are `python` in this case.
-                KeywordImpression.objects.using("default").filter(
-                    pk=impression.pk
-                ).update(**{impression_type: F(impression_type) + values["total"]})
+            # These are a Sum because we can't query for specific keywords from Postgres,
+            # so a specific publisher and advertisement set could return the same keyword:
+            # ['python', 'django'] and ['python, 'flask'] both are `python` in this case.
+            KeywordImpression.objects.using("default").filter(pk=impression.pk).update(
+                decisions=F("decisions") + values["total_decisions"],
+                offers=F("offers") + values["total_offers"],
+                views=F("views") + values["total_views"],
+                clicks=F("clicks") + values["total_clicks"],
+            )
 
 
 @app.task()
@@ -328,89 +343,104 @@ def daily_update_regiontopic(day=None):  # pylint: disable=too-many-branches
         date__gte=start_date, date__lt=end_date
     ).delete()
 
-    for impression_type in IMPRESSION_TYPES:
-        topic_mapping = defaultdict(int)
-        queryset = _default_filters(impression_type, start_date, end_date)
-        for values in (
-            queryset.values("advertisement", "keywords", "country")
-            .annotate(total=Count("country"))
-            .filter(total__gt=0)
-            .order_by("-total")
-            .values("keywords", "advertisement", "country", "total")
-        ):
-            if not (values["keywords"] and values["country"]):
+    topic_mapping = defaultdict(
+        lambda: {
+            "decisions": 0,
+            "offers": 0,
+            "views": 0,
+            "clicks": 0,
+        }
+    )
+    queryset = Offer.objects.using(settings.REPLICA_SLUG).filter(
+        date__gte=start_date,
+        date__lt=end_date,  # Things at UTC midnight should count towards tomorrow
+    )
+    for values in (
+        queryset.values("advertisement", "keywords", "country")
+        .annotate(
+            total_decisions=Count("country"),
+            total_offers=Count("country", filter=Q(advertisement__isnull=False)),
+            total_views=Count("country", filter=Q(viewed=True)),
+            total_clicks=Count("country", filter=Q(clicked=True)),
+        )
+        .filter(total_decisions__gt=0)
+        .order_by("-total_decisions")
+        .values(
+            "keywords",
+            "advertisement",
+            "country",
+            "total_decisions",
+            "total_offers",
+            "total_views",
+            "total_clicks",
+        )
+        .iterator()
+    ):
+        if not (values["keywords"] and values["country"]):
+            continue
+
+        keywords = values["keywords"]
+        country = values["country"]
+        ad = values["advertisement"]
+        publisher_keywords = set(keywords)
+
+        topics = set()
+        for keyword in publisher_keywords:
+            if keyword in data_science:
+                topic = "data-science"
+            elif keyword in backend_web:
+                topic = "backend-web"
+            elif keyword in frontend_web:
+                topic = "frontend-web"
+            elif keyword in security_privacy:
+                topic = "security-privacy"
+            elif keyword in devops:
+                topic = "devops"
+            elif keyword in python:
+                topic = "python"
+            elif keyword in blockchain:
+                topic = "blockchain"
+            elif keyword in game_dev:
+                topic = "game-dev"
+            else:
                 continue
 
-            keywords = json.loads(values["keywords"])
-            country = values["country"]
-            ad = values["advertisement"]
-            publisher_keywords = set(keywords)
+            topics.add(topic)
 
-            topics = set()
-            for keyword in publisher_keywords:
-                if keyword in data_science:
-                    topic = "data-science"
-                elif keyword in backend_web:
-                    topic = "backend-web"
-                elif keyword in frontend_web:
-                    topic = "frontend-web"
-                elif keyword in security_privacy:
-                    topic = "security-privacy"
-                elif keyword in devops:
-                    topic = "devops"
-                elif keyword in python:
-                    topic = "python"
-                elif keyword in blockchain:
-                    topic = "blockchain"
-                elif keyword in game_dev:
-                    topic = "game-dev"
-                else:
-                    continue
+        # If nothing gets set as a topic, assign it other
+        if not topics:
+            topics.add("other")
 
-                topics.add(topic)
+        region = get_region_from_country_code(country)
 
-            # If nothing gets set as a topic, assign it other
-            if not topics:
-                topics.add("other")
+        # Aggregate data into topic_mapping to save doing queries until we have everything counted
+        # This is important because we can't query on keywords, so we have a lot of records that increment
+        # the total count on the region & topic.
+        for topic in topics:
+            topic_mapping[f"{ad}:{region}:{topic}"]["decisions"] += values[
+                "total_decisions"
+            ]
+            topic_mapping[f"{ad}:{region}:{topic}"]["offers"] += values["total_offers"]
+            topic_mapping[f"{ad}:{region}:{topic}"]["views"] += values["total_views"]
+            topic_mapping[f"{ad}:{region}:{topic}"]["clicks"] += values["total_clicks"]
 
-            if country in us_ca:
-                region = "us-ca"
-            elif country in eu_aus_nz:
-                region = "eu-aus-nz"
-            elif country in wider_apac:
-                region = "wider-apac"
-            elif country in latin_america:
-                region = "latin-america"
-            elif country in africa:
-                region = "africa"
-            else:
-                region = "global"
-
-            # Aggregate data into topic_mapping to save doing queries until we have everything counted
-            # This is important because we can't query on keywords, so we have a lot of records that increment
-            # the total count on the region & topic.
-            for topic in topics:
-                topic_mapping[f"{ad}:{region}:{topic}"] += values["total"]
-
-        log.info(
-            "Saving %s RegionTopicImpressions: %s", len(topic_mapping), impression_type
+    for data, value in topic_mapping.items():
+        ad, region, topic = data.split(":")
+        # Handle the conversion of
+        if ad == "None":
+            ad = None
+        impression, _ = RegionTopicImpression.objects.using("default").get_or_create(
+            date=start_date, advertisement_id=ad, region=region, topic=topic
         )
-        for data, value in topic_mapping.items():
-            ad, region, topic = data.split(":")
-            # Handle the conversion of
-            if ad == "None":
-                ad = None
-            impression, _ = RegionTopicImpression.objects.using(
-                "default"
-            ).get_or_create(
-                date=start_date, advertisement_id=ad, region=region, topic=topic
-            )
-            # these are a sum because we can't query for specific keywords from postgres,
-            # so a specific publisher and advertisement set could return the same keyword:
-            # ['python', 'django'] and ['python, 'flask'] both are `python` in this case.
-            RegionTopicImpression.objects.using("default").filter(
-                pk=impression.pk
-            ).update(**{impression_type: F(impression_type) + value})
+        # these are a sum because we can't query for specific keywords from postgres,
+        # so a specific publisher and advertisement set could return the same keyword:
+        # ['python', 'django'] and ['python, 'flask'] both are `python` in this case.
+        RegionTopicImpression.objects.using("default").filter(pk=impression.pk).update(
+            decisions=F("decisions") + value["decisions"],
+            offers=F("offers") + value["offers"],
+            views=F("views") + value["views"],
+            clicks=F("clicks") + value["clicks"],
+        )
 
 
 @app.task()
@@ -424,25 +454,42 @@ def daily_update_uplift(day=None):
 
     log.info("Updating uplift for %s-%s", start_date, end_date)
 
-    for impression_type in IMPRESSION_TYPES:
-        queryset = _default_filters(impression_type, start_date, end_date)
+    queryset = Offer.objects.using(settings.REPLICA_SLUG).filter(
+        date__gte=start_date,
+        date__lt=end_date,  # Things at UTC midnight should count towards tomorrow
+    )
 
-        for values in (
-            queryset.values("publisher", "advertisement")
-            .annotate(total=Count("uplifted"))
-            .filter(total__gt=0)
-            .order_by("-total")
-            .values("publisher", "advertisement", "total")
-        ):
-
-            impression, _ = UpliftImpression.objects.using("default").get_or_create(
-                publisher_id=values["publisher"],
-                advertisement_id=values["advertisement"],
-                date=start_date,
-            )
-            UpliftImpression.objects.using("default").filter(pk=impression.pk).update(
-                **{impression_type: values["total"]}
-            )
+    for values in (
+        queryset.values("publisher", "advertisement")
+        .annotate(
+            total_decisions=Count("uplifted"),
+            total_offers=Count("uplifted", filter=Q(advertisement__isnull=False)),
+            total_views=Count("uplifted", filter=Q(viewed=True)),
+            total_clicks=Count("uplifted", filter=Q(clicked=True)),
+        )
+        .filter(total_decisions__gt=0)
+        .order_by("-total_decisions")
+        .values(
+            "publisher",
+            "advertisement",
+            "total_decisions",
+            "total_offers",
+            "total_views",
+            "total_clicks",
+        )
+        .iterator()
+    ):
+        impression, _ = UpliftImpression.objects.using("default").get_or_create(
+            publisher_id=values["publisher"],
+            advertisement_id=values["advertisement"],
+            date=start_date,
+        )
+        UpliftImpression.objects.using("default").filter(pk=impression.pk).update(
+            decisions=values["total_decisions"],
+            offers=values["total_offers"],
+            views=values["total_views"],
+            clicks=values["total_clicks"],
+        )
 
 
 @app.task(time_limit=60 * 60 * 4)
@@ -454,8 +501,10 @@ def update_previous_day_reports(day=None):
     """
     start_date, _ = _get_day(day)
 
-    # Do the previous day now that the day is complete
-    start_date -= datetime.timedelta(days=1)
+    if not day:
+        # If not specified,
+        # do the previous day now that the day is complete
+        start_date -= datetime.timedelta(days=1)
 
     # Do all reports
     daily_update_geos(start_date)
@@ -465,11 +514,15 @@ def update_previous_day_reports(day=None):
     daily_update_uplift(start_date)
     daily_update_regiontopic(start_date)
 
-    # Send notification to Slack about previous day's reports
-    slack_message(
-        "adserver/slack/generic-message.slack",
-        {"text": "Completed aggregating report data for yesterday. :page_with_curl:"},
-    )
+    if not day:
+        # Send notification to Slack about previous day's reports
+        # Don't send this notification if run manually
+        slack_message(
+            "adserver/slack/generic-message.slack",
+            {
+                "text": f"Completed aggregating report data for yesterday ({start_date:%Y-%m-%d}). :page_with_curl:"
+            },
+        )
 
 
 @app.task()
@@ -505,9 +558,9 @@ def calculate_publisher_ctrs(days=7):
 
 @app.task()
 def notify_of_completed_flights():
-    """Send a notification about flights which completed in the last day."""
+    """Send a note and close flights which completed in the last day."""
     cutoff = get_ad_day() - datetime.timedelta(days=1)
-    for flight in Flight.objects.filter(live=True).filter():
+    for flight in Flight.objects.filter(live=True):
         if (
             flight.clicks_remaining() == 0
             and flight.views_remaining() == 0
@@ -525,6 +578,10 @@ def notify_of_completed_flights():
                     "flight_url": generate_absolute_url(flight.get_absolute_url()),
                 },
             )
+
+            # Mark the flight as no longer live. It is complete
+            flight.live = False
+            flight.save()
 
 
 @app.task()
@@ -558,7 +615,7 @@ def notify_of_publisher_changes(difference_threshold=0.25, min_views=10_000):
         previous_week_report = PublisherReport(queryset)
         previous_week_report.generate()
 
-        for metric in ("views", "ctr"):
+        for metric in ("views",):
             total_views = (
                 last_week_report.total["views"] + previous_week_report.total["views"]
             )
