@@ -21,7 +21,7 @@ from .base import env
 
 
 # Django Settings
-# https://docs.djangoproject.com/en/2.2/ref/settings/
+# https://docs.djangoproject.com/en/3.2/ref/settings/
 # --------------------------------------------------------------------------
 DEBUG = False
 TEMPLATE_DEBUG = DEBUG
@@ -34,7 +34,7 @@ INTERNAL_IPS = env.list("INTERNAL_IPS", default=[])
 
 
 # Database
-# https://docs.djangoproject.com/en/2.2/ref/settings/#databases
+# https://docs.djangoproject.com/en/3.2/ref/settings/#databases
 # --------------------------------------------------------------------------
 DATABASES["default"] = env.db()  # Raises ImproperlyConfigured if DATABASE_URL not set
 DATABASES["default"]["CONN_MAX_AGE"] = env.int("CONN_MAX_AGE", default=3600)
@@ -50,7 +50,7 @@ LOGGING["loggers"]["django.security.DisallowedHost"] = {
 
 
 # Cache
-# https://docs.djangoproject.com/en/2.2/topics/cache/
+# https://docs.djangoproject.com/en/3.2/topics/cache/
 # https://niwinz.github.io/django-redis/
 # --------------------------------------------------------------------------
 CACHES = {"default": env.cache("REDIS_URL")}
@@ -64,9 +64,9 @@ if env.bool("REDIS_SSL", default=False):
 
 
 # Security
-# https://docs.djangoproject.com/en/2.2/topics/security/
-# https://docs.djangoproject.com/en/2.2/ref/middleware/#django.middleware.security.SecurityMiddleware
-# https://docs.djangoproject.com/en/2.2/ref/clickjacking/
+# https://docs.djangoproject.com/en/3.2/topics/security/
+# https://docs.djangoproject.com/en/3.2/ref/middleware/#django.middleware.security.SecurityMiddleware
+# https://docs.djangoproject.com/en/3.2/ref/clickjacking/
 # --------------------------------------------------------------------------
 if env.bool("ADSERVER_HTTPS", default=False):
     ADSERVER_HTTPS = True
@@ -93,7 +93,7 @@ ANYMAIL = {"SENDGRID_API_KEY": env("SENDGRID_API_KEY")}
 
 
 # User upload storage
-# https://docs.djangoproject.com/en/2.2/topics/files/
+# https://docs.djangoproject.com/en/3.2/topics/files/
 # https://django-storages.readthedocs.io/en/latest/backends/azure.html
 DEFAULT_FILE_STORAGE = env(
     "DEFAULT_FILE_STORAGE", default="storages.backends.azure_storage.AzureStorage"
@@ -184,3 +184,6 @@ ADSERVER_STICKY_DECISION_DURATION = env.int(
 GEOIP_PATH = env("GEOIP_GEOLITE2_PATH", default=GEOIP_PATH)
 GEOIP_CITY = env("GEOIP_GEOLITE2_CITY_FILENAME", default="GeoLite2-City.mmdb")
 GEOIP_COUNTRY = env("GEOIP_GEOLITE2_COUNTRY_FILENAME", default="GeoLite2-Country.mmdb")
+
+# Stripe settings for dj-stripe
+STRIPE_LIVE_MODE = True
